fix(refactoring): Slicing the codebase
