+++ resolveArgs.js
@@ -1,0 +1,49 @@
+const encodings = new Set(['json', 'buffer', 'string'])
+
+const resolveArgs = (args) => {
+  let statusCodes = new Set()
+  let method
+  let encoding
+  let headers
+  let baseurl = ''
+
+  args.forEach(arg => {
+    if (typeof arg === 'string') {
+      if (arg.toUpperCase() === arg) {
+        if (method) {
+          let msg = `Can't set method to ${arg}, already set to ${method}.`
+          throw new Error(msg)
+        } else {
+          method = arg
+        }
+      } else if (arg.startsWith('http:') || arg.startsWith('https:')) {
+        baseurl = arg
+      } else {
+        if (encodings.has(arg)) {
+          encoding = arg
+        } else {
+          throw new Error(`Unknown encoding, ${arg}`)
+        }
+      }
+    } else if (typeof arg === 'number') {
+      statusCodes.add(arg)
+    } else if (typeof arg === 'object') {
+      if (headers) {
+        throw new Error('Cannot set headers twice.')
+      }
+      headers = arg
+    } else {
+      throw new Error(`Unknown type: ${typeof arg}`)
+    }
+  })
+
+  return {
+    statusCodes,
+    method,
+    encoding,
+    headers,
+    baseurl
+  }
+}
+
+module.exports = resolveArgs
